Return nil when accessing non-current components of a "oneof"
Currently a message alternative in a "oneof" appears as an empty table when it
is not the currently active member of the oneof, whereas getting nil would be
preferred.

Update to track the currently active member of the oneof and return nil
whenever a different member is accessed.

The _setter metamethod updates the current member since this is used when
application code writes to a field and in that case switching is expected.

In line with https://developers.google.com/protocol-buffers/docs/proto3#oneof
the Decoder also updates rather than rejecting.

The _getter metamethod returns nil for any non-current member of the oneof.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>